Crate Elements_of_Programming_in_Rust

Modules

problem_4_1
EoPI pg 24 The parity of a binary word is 1 if the number of 1s in the word is odd; otherwise, it is 0. For example, the parity of 1011 is 1, and the parity of 10001000 is 0. Parity checks are used to detect single bit errors in data storage and communication. It is fairly straightforward to write code that computes the parity of a single 64-bit word. How would you compute the parity of a very large number of 64-bit words? Hint: Use a lookup table, but don't use 2^64 entries!

problem_5_1
EOPI pg 39. The Dutch National Flag Problem Write a program that takes an array slice and an index i into slice, and rearranges the elements such that all elements less than slice[r] (the "pivot") appear first, followed by elements equal to the pivot, followed by elements greater than the pivot. Hint: Think about the partition step in quicksort. the documentation for the slice primitive comes in handy

problem_5_6
EoPI pg 46 Buy and sell a stock once This problem is concerned with the problem of optimally buying and selling a stock once, as described on Page 2. As an example, consider the following sequence of stock prices: <31.0,31.5,275,295,260,270,290,230,255,250>. The maximum profit that can be made with one buy and one sell is 30-buy at 260 and sell at 290. Note that 250 is not the lowest price, nor 290 the highest price. 46Write program that takes an array denoting the daily stock price, and retums the maximum profit that could be made by buying and then selling one share of that stock. There is no need to buy if a no profit is possible. Hint:ldentifying the minimum and maximum height. Focus on valid differences.

problem_6_1
EoPI pg 68 Interconvert Strings and Integers In this problem, you are to irnplement methods that take a string representing an integer and retum the corresponding integer, and vice versa. Your code should handle negative integers. You cannot use library functions like int in Python. Implement an integer to string conversion function, and a string to integer conversison function, For example, if the input to the first function is the integer 314,it should retum the string "31.4" and if the input to the second function is the string "314" it should return the integer 314. Hint: Build the result one digit at a time.

problem_7_1
7.1 Merge two sorted lists Consider two singly linked lists in which each node holds a number. Assume the lists are sorted, i.e., numbers in the lists appear in ascending order within each list. The merge of the two lists is a list consisting of the nodes of the two lists in which numbers appear in ascending order

problem_8_1
8.1 Implement a stack with a max api Design a stack that includes a max operation, in addition to push and pop. The max method should return the maximum value stored in the stack.

utils