Module Elements_of_Programming_in_Rust::problem_4_1

EoPI pg 24 The parity of a binary word is 1 if the number of 1s in the word is odd; otherwise, it is 0. For example, the parity of 1011 is 1, and the parity of 10001000 is 0. Parity checks are used to detect single bit errors in data storage and communication. It is fairly straightforward to write code that computes the parity of a single 64-bit word. How would you compute the parity of a very large number of 64-bit words? Hint: Use a lookup table, but don't use 2^64 entries!

Structs

Solution1
recursively compute the parity of right shifted subwords NO dynamic prog lookup in hashmap

Solution2
iteratively fold a word in half XOR'ing all the way NO dynamic prog lookup in hashmap

Solution3
iteratively fold a word in half XOR'ing all the way YES dynamic prog lookup in hashmap

Traits

Parity
Fundamental Trait we will be implementing in this problem.

Functions

n_bits
A helper function for getting the number of bits of a type T.