Module Elements_of_Programming_in_Rust::problem_5_6

EoPI pg 46 Buy and sell a stock once This problem is concerned with the problem of optimally buying and selling a stock once, as described on Page 2. As an example, consider the following sequence of stock prices: <31.0,31.5,275,295,260,270,290,230,255,250>. The maximum profit that can be made with one buy and one sell is 30-buy at 260 and sell at 290. Note that 250 is not the lowest price, nor 290 the highest price. 46Write program that takes an array denoting the daily stock price, and retums the maximum profit that could be made by buying and then selling one share of that stock. There is no need to buy if a no profit is possible. Hint:ldentifying the minimum and maximum height. Focus on valid differences.

Functions

buy_sell
the primary solution takes a slice &[T] and returns a T where references to T can be subtracted (impl Sub for &T {type Output = T}) and are orderable (impl Ord for T) the solution iterates through the slice once and for each element iterates through the remainder of the slice for each element of the inner loop it find the difference which is the amount gained from selling if a better result is found than what is currently had then that becomes the new solution