Module Elements_of_Programming_in_Rust::utils::smart_pointers::rc

Rc is a type that implements reference counting at runtime it's purpose is, unlike Cell and RefCell, not to supply interior mutability but to facilitate sharing a value in multiple places when lifetimes are not certain an Rc<T> does not give out means by which to mutate the T only ones to get reference to it

Structs

Rc
the actual Rc object

RcInner
RcInner is what an Rc actually stores a reference to. it is necessary in a similar but not the same way that Ref and RefMut are necessary to RefCell when an Rc is cloned we need a way to continue tracking the refcount and so we do this by storing a raw pointer to RcInner and when the Rc is cloned the raw pointer to the RcInner is cloned too