Module Elements_of_Programming_in_Rust::utils::smart_pointers::refcell

RefCell is a type that provides interior mutability at runtime for types that are less efficiently Copy

Structs

Ref
the Ref object that will be used to return references with an api for semantic Deref and Drop see impl Drop for Ref for more explanation

RefCell
The Refcell itself

RefMut
Refmut is analagous to 'Refbut is returned byRefcell's borrow_mutit will allow dereferencing to a mutable reference to theRefcell`'s internal value

Enums

BorrowStatus
BorrowStatus represents the state of a RefCell at runtime